// src/app/models/event-interface.ts
export interface EventInterface {
    id?: string;
    name?: string;
    slug?: string;
    date?: string;
    time?: string;
    address?: string;
    quotas?: string;
    expositor?: string;
    addressed?: string;
    description?: string;
    poster?: string;
    banner?: string;
    state?: string;
    category?: string;
    created_at?: string;
    updated_at?: string;
    data?: any;

    
}